platform/learning/cross_tenant_store.py
from models.benchmark import BenchmarkResult
from models.cluster import ClusterProfile
from models.optimization import OptimizationBundle
import psycopg2
import psycopg2.extras

def store_anonymized_benchmark(benchmark):
    # Placeholder for storing anonymized benchmark
    print(f"Storing anonymized benchmark: {benchmark}")

def get_pre_warmed_recommendations(cluster_profile):
    # Placeholder for getting pre-warmed recommendations
    return [{"model_sha256": "model_sha256_1", "family": "llama", "size": 8, "quant": "Q8", "backend": "llama.cpp", "decode_tps": 120}]