platform/requirements.txt
fastapi
uvicorn
python-jose
asyncio
aiohttp
psycopg2-binary
celery
redis
stripe
pydantic
alembic
pgvector
chromadb
